feat: Adds google login
This commit is a big one. Adding google oauth sign-in needs
a lot of work: gconnect/gdisconnect functions in ptoject.py,
login.html template and credential file.

+++ vagrant/catalog/project.py
@@ -17,6 +17,11 @@ from sqlalchemy.orm import sessionmaker
 #  Creates flask app
 app = Flask(__name__)
 
+CLIENT_ID = json.loads(
+    open('client_secrets.json', 'r').read())['web']['client_id']
+FB_APP_ID = json.loads(
+    open('fb_client_secrets.json', 'r').read())['web']['app_id']
+
 #  Connects to PostgreSQL and creates a database session
 engine = create_engine('postgresql:///catalog')
 Base.metadata.bind = engine
@@ -40,7 +45,11 @@ def showLogin():
     state = ''.join(random.choice(string.ascii_uppercase + string.digits)
                     for x in xrange(32))
     login_session['state'] = state
-    return render_template('login.html', state=state, on_login=True)
+    return render_template('login.html',
+                           state=state,
+                           fb_app_id=FB_APP_ID,
+                           client_id=CLIENT_ID,
+                           on_login=True)
 
 
 @app.route('/fbconnect', methods=['POST'])
@@ -123,6 +132,119 @@ def fbdisconnect():
     return "you have been logged out"
 
 
+@app.route('/gconnect', methods=['POST'])
+def gconnect():
+    # Validate state token
+    if request.args.get('state') != login_session['state']:
+        response = make_response(json.dumps('Invalid state parameter.'), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+    # Obtain authorization code
+    code = request.data
+
+    try:
+        # Upgrade the authorization code into a credentials object
+        oauth_flow = flow_from_clientsecrets('client_secrets.json', scope='')
+        oauth_flow.redirect_uri = 'postmessage'
+        credentials = oauth_flow.step2_exchange(code)
+    except FlowExchangeError:
+        response = make_response(
+            json.dumps('Failed to upgrade the authorization code.'), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+    # Check that the access token is valid.
+    access_token = credentials.access_token
+    url = ('https://www.googleapis.com/oauth2/v1/tokeninfo?access_token=%s'
+           % access_token)
+    h = httplib2.Http()
+    result = json.loads(h.request(url, 'GET')[1])
+
+    # If there was an error in the access token info, abort.
+    if result.get('error') is not None:
+        response = make_response(json.dumps(result.get('error')), 500)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+    # Verify that the access token is used for the intended user.
+    gplus_id = credentials.id_token['sub']
+    if result['user_id'] != gplus_id:
+        response = make_response(
+            json.dumps("Token's user ID doesn't match given user ID."), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+    # Verify that the access token is valid for this app.
+    if result['issued_to'] != CLIENT_ID:
+        response = make_response(
+            json.dumps("Token's client ID does not match app's."), 401)
+        print "Token's client ID does not match app's."
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+    stored_credentials = login_session.get('credentials')
+    stored_gplus_id = login_session.get('gplus_id')
+    if stored_credentials is not None and gplus_id == stored_gplus_id:
+        response = make_response(json.dumps('Current user is already connected.'),
+                                 200)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+    # Store the access token in the session for later use.
+    login_session['access_token'] = credentials.access_token
+    login_session['gplus_id'] = gplus_id
+
+    # Get user info
+    userinfo_url = "https://www.googleapis.com/oauth2/v1/userinfo"
+    params = {'access_token': credentials.access_token, 'alt': 'json'}
+    answer = requests.get(userinfo_url, params=params)
+
+    data = answer.json()
+
+    login_session['username'] = data['name']
+    login_session['picture'] = data['picture']
+    login_session['email'] = data['email']
+    # ADD PROVIDER TO LOGIN SESSION
+    login_session['provider'] = 'google'
+
+    # see if user exists, if it doesn't make a new one
+    user_id = getUserID(data["email"])
+    if not user_id:
+        user_id = createUser(login_session)
+    login_session['user_id'] = user_id
+
+    output = ''
+    output += '<h1>Welcome, '
+    output += login_session['username']
+    output += '!</h1>'
+    output += '<img src="'
+    output += login_session['picture']
+    output += ' " style = "width: 300px; height: 300px;border-radius: 150px;-webkit-border-radius: 150px;-moz-border-radius: 150px;"> '
+    flash("you are now logged in as %s" % login_session['username'])
+    print "done!"
+    return output
+
+@app.route('/gdisconnect')
+def gdisconnect():
+    # Only disconnect a connected user.
+    credentials = login_session.get('credentials')
+    if credentials is None:
+        response = make_response(
+            json.dumps('Current user not connected.'), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+    access_token = credentials.access_token
+    url = 'https://accounts.google.com/o/oauth2/revoke?token=%s' % access_token
+    h = httplib2.Http()
+    result = h.request(url, 'GET')[0]
+    if result['status'] != '200':
+        # For whatever reason, the given token was invalid.
+        response = make_response(
+            json.dumps('Failed to revoke token for given user.'), 400)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+
 # User Helper Functions
 def createUser(login_session):
     """Helper function to create a user"""
@@ -157,6 +279,10 @@ def disconnect():
         if login_session['provider'] == 'facebook':
             fbdisconnect()
             del login_session['facebook_id']
+        elif login_session['provider'] == 'google':
+            gdisconnect()
+            del login_session['gplus_id']
+            del login_session['access_token']
         del login_session['username']
         del login_session['email']
         del login_session['picture']
@@ -166,7 +292,7 @@ def disconnect():
         return redirect(url_for('showCategoriesBooks'))
     else:
         flash("You were not logged in")
-        return redirect(url_for('showCtegoriesBooks'))
+        return redirect(url_for('showCategoriesBooks'))
 
 
 @app.route('/')
@@ -187,13 +313,17 @@ def showCategoriesBooks():
 def showBooksCategory(category_id):
     """Renders all books of a specified category id"""
     categories = session.query(Category).all()
-    category = session.query(Category).filter_by(id=category_id).one()
+    category = session.query(Category).filter_by(id=category_id).one_or_none()
+    if not category:
+        flash("No category found (id=%s)" % category_id)
+        return redirect(url_for('showCategoriesBooks'))
+
     books = session.query(Book).filter_by(category_id=category_id).order_by(Book.id.desc())
     logged_in = True if 'username' in login_session else False
     return render_template('book_catalog.html',
                            logged_in=logged_in,
                            categories=categories,
-                           category_name=category.name,
+                           category_name=category.name if category else None,
                            number_of_books=books.count(),
                            books=books)
 
@@ -201,7 +331,11 @@ def showBooksCategory(category_id):
 @app.route('/book_catalog/books/<int:book_id>')
 def showBookDescription(book_id):
     """Renders a description of specified book id"""
-    book = session.query(Book).filter_by(id=book_id).one()
+    book = session.query(Book).filter_by(id=book_id).one_or_none()
+    if not book:
+        flash("No book found (id=%s)" % book_id)
+        return redirect(url_for('showCategoriesBooks'))
+
     logged_in = login_session['username'] if 'username' in login_session else None
     return render_template('book_description.html',
                            logged_in=logged_in,
@@ -214,9 +348,9 @@ def bookData(request):
     """
     if request.form.get('category'):
         category = session.query(Category).filter_by(
-            name=request.form.get('category')).one()
+            name=request.form.get('category')).one_or_none()
     else:
-        categoty = None
+        category = None
     data = {
         'title': request.form.get('title'),
         'author': request.form.get('author'),
@@ -310,8 +444,8 @@ def editBook(book_id):
                                    categories=categories,
                                    data=data)
         else:
-            book = session.query(Book).filter_by(id=book_id).one()
-            if book.user_id == login_session['user_id']:
+            book = session.query(Book).filter_by(id=book_id).one_or_none()
+            if book and (book.user_id == login_session['user_id']):
                 if 'title' in data: book.title = data['title']
                 if 'author' in data: book.author = data['author']
                 if 'description' in data: book.description = data['description']
@@ -326,21 +460,22 @@ def editBook(book_id):
                       % (logged_in, book.title))
             return redirect(url_for('showCategoriesBooks'))
     else:
-        book = session.query(Book).filter_by(id=book_id).one()
+        book = session.query(Book).filter_by(id=book_id).one_or_none()
         categories = session.query(Category).all()
-        data = {
-            'title': book.title,
-            'author': book.author if book.author else '',
-            'category': book.category.name if book.category else '',
-            'price': book.price if book.price else '$ 0.00',
-            'description': book.description if book.description else ''
-        }
+        if book:
+            data = {
+                'title': book.title,
+                'author': book.author if book.author else '',
+                'category': book.category.name if book.category else '',
+                'price': book.price if book.price else '$ 0.00',
+                'description': book.description if book.description else ''
+            }
         return render_template('book_edit.html',
                                logged_in=logged_in,
                                title='Edit Book',
                                button='Update',
                                categories=categories,
-                               data=data)
+                               data=data if data else {})
 
 
 @app.route('/book_catalog/books/<int:book_id>/delete', methods=['GET', 'POST'])
@@ -351,8 +486,8 @@ def deleteBook(book_id):
         flash('Login to delete a book you have created.')
         return redirect(url_for('showCategoriesBooks'))
 
-    book = session.query(Book).filter_by(id=book_id).one()
-    if request.method == 'POST':
+    book = session.query(Book).filter_by(id=book_id).one_or_none()
+    if book and (request.method == 'POST'):
         if book.user_id == login_session['user_id']:
             session.delete(book)
             session.commit()
@@ -368,5 +503,4 @@ def deleteBook(book_id):
 
 if __name__ == '__main__':
     app.secret_key = 'super_secret_key'
-    app.debug = True
     app.run(host='0.0.0.0', port=5000)
